# Improve SHA/HMAC performance