--- a/openssl/goopenssl.h
+++ b/openssl/goopenssl.h
@@ -41,6 +41,7 @@ typedef void* GO_EVP_PKEY_PTR;
 typedef void* GO_EVP_PKEY_CTX_PTR;
 typedef void* GO_EVP_MD_PTR;
 typedef void* GO_EVP_MD_CTX_PTR;
+typedef void* GO_HMAC_CTX_PTR;
 
 // Define pointers to all the used OpenSSL functions.
 // Calling C function pointers from Go is currently not supported.