--- a/openssl/hmac.go
+++ b/openssl/hmac.go
@@ -61,7 +61,7 @@ type opensslHMAC struct {
 func (h *opensslHMAC) Reset() {
 	hmacCtxReset(h.ctx)
 
-	if C.go_openssl_HMAC_Init_ex(h.ctx, unsafe.Pointer(base(h.key)), C.int(len(h.key)), h.md, nil) == 0 {
+	if C.go_openssl_HMAC_Init_ex(h.ctx, base(h.key), C.int(len(h.key)), h.md, nil) == 0 {
 		panic("openssl: HMAC_Init failed")
 	}
 	if size := C.go_openssl_EVP_MD_get_size(h.md); size != C.size_t(h.size) {