--- a/openssl/sha_test.go
+++ b/openssl/sha_test.go
@@ -70,7 +70,7 @@ func BenchmarkHash8Bytes(b *testing.B) {
 	b.StopTimer()
 	h := NewSHA256()
 	sum := make([]byte, h.Size())
-	var buf = make([]byte, 8192)
+	buf := make([]byte, 8192)
 	size := 1024
 	b.StartTimer()
 	b.SetBytes(int64(size))