SHA/HMAC throughput can't be significantly improved because the bottleneck is in the OpenSSL cgo call that performs the crypto heavy lifting, namely EVP_DigestUpdate and HMAC_Update. These accounts for ~90% of the processing time.

Batching cgo calls nor reducing GC pressure help increase the throughput because we are at the order of 3.000 ns/op, much higher than the ~60ns cgo overhead and whatever the GC latency has.

On the other hand, we can reduce the allocations by defining `EVP_MD`, `EVP_MD_CTX` and `HMAC_CTX` pointer types in C instead of Go. This trick avoids filling Go heap with unnecessary pointers unless strictly necessary.

The result is that we have a tiny improved throughput, and we have reduced allocated memory and objects.

```txt
name               old time/op    new time/op    delta
Hash8Bytes-4         3.67µs ± 5%    3.50µs ± 6%    -4.77%  (p=0.018 n=9+10)
HMACSHA256_32-4      2.39µs ± 3%    2.28µs ± 4%    -4.71%  (p=0.000 n=9+10)
HMACNewWriteSum-4    5.14µs ± 4%    4.85µs ± 7%    -5.62%  (p=0.001 n=10+9)

name               old speed      new speed      delta
Hash8Bytes-4        279MB/s ± 5%   293MB/s ± 5%    +5.08%  (p=0.018 n=9+10)
HMACSHA256_32-4    13.4MB/s ± 3%  14.0MB/s ± 4%    +4.97%  (p=0.000 n=9+10)
HMACNewWriteSum-4  6.23MB/s ± 4%  6.61MB/s ± 7%    +6.01%  (p=0.001 n=10+9)

name               old B/op   new B/op   delta
Hash8Bytes-4           104B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
HMACSHA256_32-4        152B ± 0%       80B ± 0%   -47.37%  (p=0.000 n=10+10)
HMACNewWriteSum-4      400B ± 0%      288B ± 0%   -28.00%  (p=0.000 n=9+10)

name               old allocs/op  new allocs/op  delta
Hash8Bytes-4           11.0 ± 0%       0.0       -100.00%  (p=0.000 n=10+10)
HMACSHA256_32-4        12.0 ± 0%       3.0 ± 0%   -75.00%  (p=0.000 n=10+10)
HMACNewWriteSum-4      21.0 ± 0%       7.0 ± 0%   -66.67%  (p=0.000 n=9+10)
```